pyproject.toml
[tool.poetry]
name = "post-ost"
version = "0.1.0"
description = ""
authors = ["Loopeq <arsen.00z11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
pydantic-settings = "^2.6.1"
uvicorn = "^0.32.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
orjson = "^3.10.11"
pydantic = {extras = ["email"], version = "^2.9.2"}
python-multipart = "^0.0.17"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiosqlite = "^0.20.0"
aiofiles = "^24.1.0"
pytest = "^8.3.3"
faker = "^33.0.0"
pytest-mock = "^3.14.0"
httpx = "^0.27.2"
psycopg2 = "^2.9.10"
bcrypt = "<4.0.0"
vectorbt = "^0.26.2"
numpy = "1.23.5"
yfinance = "^0.2.50"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
force-exclude = '''
/(
 | src/alembic
 | src/strategies/root
)/
'''

line-length = 80

